Cambio de long a int

+++ AppAdminSIE_BE/AppAdminSIE_BE/Data/Interfaces/IUsuarioXServicioRepository.cs
@@ -5,7 +5,7 @@ namespace JobOclock_BackEnd.Data.Interfaces
     public interface IUsuarioXServicioRepository
     {
         IEnumerable<UsuarioXServicio> GetByUsuario(int idUsuario);
-        long Add(UsuarioXServicio registro);
+        int Add(UsuarioXServicio registro);
         void UpdateEstado(int idUsuario, int idActividad, string nuevoEstado);
     }
 }
